refactor: remove fluent builder pattern from eventbus
Remove SubscriptionBuilder class and related fluent API methods
to simplify the subscription architecture. Move action handler
factories to dedicated abstractions module for better organization.

- Delete SubscriptionBuilder.cs file
- Remove Builder() method from ISubscriptionManager interface
- Remove Builder() implementation from SubscriptionManager
- Move ActionEventHandlerFactory to Abstractions/Handlers module

(eventbus-architecture)
